pyproject.toml
[project]
name = "gemini-enterprise-nexus-agent"
version = "0.1.3"
description = "Nexus Intelligence Hub Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.26.0",
]

[build-system]
requires = ["uv_build>=0.9.10,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
